refactor: `PartOfSpeechDetail` を移動
